Revert "rename useCanvasSigner file"
This reverts commit dfd148674d0a9abe5a26f4ae194afdaf5459aeb7.
